temp: remove let bindings
If we have local equations which are not stateful we create a let
binding so we can remove the variables at the level of an smt leaf. We
don't want to do this if we are slicing because now we will have let
bindings and equations. For now we disable the generation of let
bindings.

This broke the regression test `frame_block_mult_assign3.lus` for some
reason.